+++ packages/jest-expo/tests/bar.js
@@ -1,0 +1,1 @@
+export const OS = ['web'];
