+++ packages/jest-expo/tests/bar.native.js
@@ -1,0 +1,1 @@
+export const OS = ['ios', 'android'];
