+++ packages/jest-expo/tests/foo.android.js
@@ -1,0 +1,1 @@
+export const OS = 'android';
