+++ packages/jest-expo/tests/foo.ios.js
@@ -1,0 +1,1 @@
+export const OS = 'ios';
