+++ packages/jest-expo/tests/foo.web.js
@@ -1,0 +1,1 @@
+export const OS = 'web';
